Drop mocks
